remove soundmanager2 and open-source the project
soundmanager2 (from 2008) was outdated and using flash/shockwave. It
was not working in modern browsers, of course. So for now we just
stubbed out the playSound() function.

William gave permission by phone to open-source the project, so we
updated both LICENSE.md and README.md to reflect the MIT license.