play win.mp3 before navigating to next game state
Previously, the win sound was played "on page load" for the next game.

This worked in 2009 when using flash-based sound. But in 2025, using the
AudioContext, the browser refuses to play sound until the user interacts
with the page. So in 2025 the win.mp3 sound simply did not play.

Now, we play the win sound before navigating to the next game state.

In addition to making the sound actually play, it seems kind of nice
to freeze the final winning moment for a bit while the sound plays,
before navigating.